Feature/docker

+++ media/application/src/main/kotlin/com/advisor/api/media/application/CreateMediaService.kt
@@ -48,6 +48,6 @@ class CreateMediaService(
             fileSize = command.fileSize
         )
 
-        return Media.create(MediaId(snowFlakeIdUtil.generateId().toLong()), mediaProps)
+        return Media.create(MediaId(snowFlakeIdUtil.generateId()), mediaProps)
     }
 }
